fix(platform): unify Linux clipboard MIME mapping so Wayland translates platform atoms
